Enemy prefab + enemy script + main menu + camera follow
Enemy Prefabs: made bat and eye prefabs that can be copied and used in large numbers

Enemy scripts: Wrote basic enemy code that can be inherited by a large number of different monsters.

Main Menu: Made play and quit buttons and some functions including keyboard/joystick selection. As well as entering and exiting the game normally, the exit function needs to be compiled into an exe file to work properly.

Camera Follow: The camera now follows the player's movement and can be self-positioned in the unity interface to smooth out the range of movement.