Rearranged the tiles for tilemapping.
The previous tile sprite was too crowded. I spread the tiles out so each set of tiles can be selected without also selecting another set of tiles.
The new tile also takes up the entire block instead of a partial block, which is more consistent in the overall design of the platforms relative to the map.