Major Revamps. Improved Spawner and CPU Movement.
Revamped Monster to be spawned instead of being in the scene from the start. Which leads to the implementation of ScriptableObjects. Now monsters can be spawned (via instantiate) anywhere on the map.
Major improvements to CPU Movement. The right and left boundaries are also spawned simultaneously with their respective monsters. What this means is that each monster has its own boundaries and will not have conflicts with each other.
At the end of each movement, the monster will idle for a random amount of time, that can be set manually in the inspector.
This will allow the CPU movement to be more realistic and natural.