Debugged Wall Jump, Wall Slide, Dash, and Monster Detection.
Wall Jump can be executed by pressing jump again when facing the walls or facing the opposite direction - must be quick for the latter. This means Coyote Time is required to make jumps and wall jumps to be more enjoyable. Coyote Time is not yet implemented at this time. It's going to be in the next commit.

Wall Slide is no longer automatic. The user can execute the wall slide when pressing the A and D keys against the walls.

Dash can now be executed while grounded. The hotkey is also changed to Left-Shift.

Monster Detection is revamped to allow larger monsters to detect the player below their center position and up to 2x the monster's height.