Debugged Traps
Fixed bugs where the player gets stuck when a trap collider is instantiated once triggered. Removed the unnecessary colliders. Resized the collider to be good enough for detecting collision and not impacting the player's input.
Revamped the delay/cooldown mechanics to be simpler.
Optimized pivot and pixel per unit for the traps.