Added BGM and adjusted the HP and XP bar.
Added a BGM_Manager script that manage the BGM during the game. If the BGM is not playing, it will play it. If it's already playing, it won't restart it. The BGM is a singleton instance like stats, and it can't be destroyed.

Changed the positions of the HP and XP bar. The HP bar is located at the top left. The XP bar stretch along the bottom of the screen.